requirements.txt
amqp==5.0.9
asgiref==3.4.1
attrs==21.4.0
billiard==3.6.4.0
black==21.12b0
celery==5.2.3
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.10
django-filter==21.1
django-reviews==1.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
dparse==0.5.1
filelock==3.4.2
flake8==4.0.1
identify==2.4.4
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
kombu==5.2.3
Markdown==3.3.6
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.22.1
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.16.0
prompt-toolkit==3.0.24
psycopg2==2.9.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
pytz==2021.3
PyYAML==6.0
redis==4.1.0
requests==2.27.1
safety==1.10.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.8
vine==5.0.0
virtualenv==20.13.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
